quasi create tutti gli effetti delle carte gialle

+++ src/main/java/it/polimi/ingsw/ps31/GameThings/ResourceList.java
@@ -20,7 +20,7 @@ public class ResourceList {
     public void setCoin() {
         System.out.println("Inserisci il valore delle monete: ");
         Scanner scanner =new Scanner(System.in);
-        int value= scanner.nextByte();
+        int value= scanner.nextInt();
         Coin coin= new Coin(0);
         coin.addValue(value);
         if (coin.getValue()!=0) {
@@ -30,7 +30,7 @@ public class ResourceList {
     public void setWood() {
         System.out.println("Inserisci il valore dei legni: ");
         Scanner scanner =new Scanner(System.in);
-        int value= scanner.nextByte();
+        int value= scanner.nextInt();
         Wood wood= new Wood(0);
         wood.addValue(value);
         if (wood.getValue()!=0) {
@@ -40,7 +40,7 @@ public class ResourceList {
     public void setStone() {
         System.out.println("Inserisci il valore delle pietre: ");
         Scanner scanner =new Scanner(System.in);
-        int value= scanner.nextByte();
+        int value= scanner.nextInt();
         Stone stone= new Stone(0);
         stone.addValue(value);
         if (stone.getValue()!=0) {
@@ -50,7 +50,7 @@ public class ResourceList {
     public void setServant() {
         System.out.println("Inserisci il valore dei servitori: ");
         Scanner scanner =new Scanner(System.in);
-        int value= scanner.nextByte();
+        int value= scanner.nextInt();
         Servant servant= new Servant(0);
         servant.addValue(value);
         if (servant.getValue()!=0) {
@@ -60,10 +60,10 @@ public class ResourceList {
     public void setMilitaryStrength() {
         System.out.println("Inserisci il valore normale dei punti militari: ");
         Scanner scanner =new Scanner(System.in);
-        int value= scanner.nextByte();
+        int value= scanner.nextInt();
         System.out.println("Inserisci il valore minimo di punti militari richiesti : ");
         Scanner scanner1 = new Scanner(System.in);
-        int valueRequest = scanner1.nextByte();
+        int valueRequest = scanner1.nextInt();
         MilitaryStrength militaryStrength= new MilitaryStrength(0,0);
         militaryStrength.addValue(value);
         militaryStrength.setValueRequest(valueRequest);
@@ -74,7 +74,7 @@ public class ResourceList {
     public void setFaithPoint() {
         System.out.println("Inserisci il valore dei punti fede: ");
         Scanner scanner =new Scanner(System.in);
-        int value= scanner.nextByte();
+        int value= scanner.nextInt();
         FaithPoint faithPoint= new FaithPoint(0);
         faithPoint.addValue(value);
         if (faithPoint.getValue()!=0) {
@@ -84,7 +84,7 @@ public class ResourceList {
     public void setVictoryPoint() {
         System.out.println("Inserisci il valore dei punti vittoria: ");
         Scanner scanner =new Scanner(System.in);
-        int value= scanner.nextByte();
+        int value= scanner.nextInt();
         VictoryPoint victoryPoint= new VictoryPoint(0);
         victoryPoint.addValue(value);
         if (victoryPoint.getValue()!=0) {
@@ -94,7 +94,7 @@ public class ResourceList {
     public void setCouncilPrivilege() {
         System.out.println("Inserisci il numero dei privilegi del consiglio: ");
         Scanner scanner =new Scanner(System.in);
-        int value= scanner.nextByte();
+        int value= scanner.nextInt();
         System.out.println("sono diversi?: ");
         boolean different= scanner.nextBoolean();
         CouncilPrivilege councilPrivilege= new CouncilPrivilege(0, false);
